Tested Permission to changes according to Roles

+++ src/main/java/com/sms/controller/TeacherController.java
@@ -1,13 +1,23 @@
 package com.sms.controller;
 
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import com.sms.model.Teacher;
 import com.sms.service.TeacherService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
+import com.sms.service.UserService;
 
 @RestController
 @RequestMapping("/api/v1")
@@ -16,6 +26,9 @@ public class TeacherController {
 
     @Autowired
     private TeacherService teacherService;
+
+    @Autowired
+    private UserService userService;
 
     // Get all the teachers
     @GetMapping("/teachers")
@@ -37,33 +50,37 @@ public class TeacherController {
 
     // Adding a new Teacher
     @PostMapping("/teacher")
-    public Teacher createTeacher(@RequestBody Teacher teacher) {
-        return teacherService.addTeacher(teacher);
+    public ResponseEntity<?> createTeacher(@RequestBody Teacher teacher) {
+        if (userService.isAdmin()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Permission denied");
+        }
+        return ResponseEntity.ok(teacherService.addTeacher(teacher));
     }
 
     // Update existing Teachers' details
     @PutMapping("/teacher/{id}")
     public ResponseEntity<?> updateTeacher(@PathVariable Long id, @RequestBody Teacher teacher) {
+        if (userService.isAdmin()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Permission denied");
+        }
         if (!id.equals(teacher.getTeacherId())) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(id + " Teacher not found");
         }
-        // Getting Updated Teacher details
         Teacher updatedTeacher = teacherService.updateTeacher(teacher);
-
-        //  and give a response
         if (updatedTeacher != null) {
             return ResponseEntity.ok(updatedTeacher);
         } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Student with ID " + id + " not found");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Teacher with ID " + id + " not found");
         }
     }
 
     // Delete teacher by ID
     @DeleteMapping("/teacher/{id}")
     public ResponseEntity<?> deleteTeacherById(@PathVariable Long id) {
+        if (userService.isAdmin()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Permission denied");
+        }
         Teacher teacher = teacherService.getTeacherById(id);
-
-        // Check the ID related teacher is in the database or not and delete
         if (teacher != null) {
             teacherService.deleteTeacher(id);
             return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Teacher with ID " + id + " deleted successfully");
